requirements.txt
asgiref==3.8.1
Django==5.0.7
django-allauth==0.63.6
django-allauth-ui==1.2.0
django-multiselectfield==0.1.13
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
python-decouple==3.8
PyYAML==6.0.1
six==1.16.0
slippers==0.6.2
sqlparse==0.5.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.1
